requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.4
    # via openstacksdk
certifi==2020.12.5
    # via
    #   geventhttpclient
    #   requests
cffi==1.14.4
    # via cryptography
chardet==4.0.0
    # via requests
click==7.1.2
    # via flask
configargparse==1.2.3
    # via locust
cryptography==3.3.1
    # via openstacksdk
decorator==4.4.2
    # via
    #   dogpile.cache
    #   openstacksdk
dogpile.cache==1.1.1
    # via openstacksdk
flask-basicauth==0.2.0
    # via locust
flask==1.1.2
    # via
    #   flask-basicauth
    #   locust
gevent==20.12.1
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==1.4.4
    # via locust
greenlet==0.4.17
    # via gevent
idna==2.10
    # via requests
iso8601==0.1.13
    # via
    #   keystoneauth1
    #   openstacksdk
itsdangerous==1.1.0
    # via flask
jinja2==2.11.2
    # via flask
jmespath==0.10.0
    # via openstacksdk
jsonpatch==1.28
    # via openstacksdk
jsonpointer==2.0
    # via jsonpatch
keystoneauth1==4.3.0
    # via openstacksdk
locust==1.4.1
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
msgpack==1.0.2
    # via locust
munch==2.5.0
    # via openstacksdk
netifaces==0.10.9
    # via openstacksdk
openstacksdk>=0.52.0
    # via -r requirements.in
os-service-types==1.7.0
    # via
    #   keystoneauth1
    #   openstacksdk
pbr==5.5.1
    # via
    #   keystoneauth1
    #   openstacksdk
    #   os-service-types
    #   stevedore
psutil==5.8.0
    # via locust
pycparser==2.20
    # via cffi
pyyaml==5.3.1
    # via openstacksdk
pyzmq==20.0.0
    # via locust
requests==2.25.1
    # via
    #   keystoneauth1
    #   locust
requestsexceptions==1.4.0
    # via openstacksdk
six==1.15.0
    # via
    #   cryptography
    #   geventhttpclient
    #   keystoneauth1
    #   munch
stevedore==3.3.0
    # via
    #   dogpile.cache
    #   keystoneauth1
urllib3==1.26.2
    # via requests
werkzeug==1.0.1
    # via
    #   flask
    #   locust
zope.event==4.5.0
    # via gevent
zope.interface==5.2.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
